+++ CHANGES.rst
@@ -14,9 +14,9 @@ CHANGES
 
 - Support task attribute for StreamResponse #1410
 
--
+- Has functional tests for client proxy #1218
 
--
+- Fix bugs with client proxy target path and proxy host with port #1413
 
 -