+++ CHANGES.rst
@@ -1,6 +1,12 @@
 CHANGES
 =======
 
+1.1.6 (2016-11-28)
+------------------
+
+- Fix `BodyPartReader.read_chunk` bug about returns zero bytes before
+  `EOF` #1428
+
 1.1.5 (2016-11-16)
 ------------------